[build] Update a bunch of "useless casts", make compiler happy
* Note, this commit doesn't fix the build entirely for GCC
dependencies still error. Even when build is successful, topaz_game crashes on login.
Please scrutinize this commit to the furthest extent possible.

+++ src/common/sql_prepared_stmt.cpp
@@ -12,7 +12,7 @@ SqlPreparedStatement::SqlPreparedStatement(MYSQL* handle, std::string const& que
         std::exit(-1);
     }
 
-    auto res = mysql_stmt_prepare(m_Statement, query.c_str(), static_cast<unsigned long>(query.size()));
+    auto res = mysql_stmt_prepare(m_Statement, query.c_str(), query.size());
     if (res != 0)
     {
         ShowError("mysql_stmt_prepare: %s", mysql_stmt_error(m_Statement));
